Allow folders as source_uris
