+++ gfw_pixetl/pixetl_prep.py
@@ -1,63 +1,13 @@
-from typing import Dict, List, Sequence, Tuple
+from typing import List, Tuple
 from urllib.parse import urlparse
 
 import click
-from google.auth.exceptions import DefaultCredentialsError
-from google.cloud import storage
-from retrying import retry
 
-from gfw_pixetl.errors import MissingGCSKeyError, retry_if_missing_gcs_key_error
 from gfw_pixetl.sources import RasterSource
 from gfw_pixetl.utils import get_bucket, upload_geometries
-from gfw_pixetl.utils.aws import get_s3_client
-
-
-class DummyTile(object):
-    """A dummy tile."""
-
-    def __init__(self, dst: Dict) -> None:
-        self.dst: Dict = dst
-        self.metadata: Dict = {}
-
-
-def get_aws_files(
-    bucket: str, prefix: str, extensions: Sequence[str] = (".tif",)
-) -> List[str]:
-    """Get all geotiffs in S3."""
-    s3_client = get_s3_client()
-    response = s3_client.list_objects_v2(Bucket=bucket, Prefix=prefix)
-
-    objs = response.get("Contents", [])
-    files = [
-        f"/vsis3/{bucket}/{obj['Key']}"
-        for obj in objs
-        if any(obj["Key"].endswith(ext) for ext in extensions)
-    ]
-
-    return files
-
-
-@retry(
-    retry_on_exception=retry_if_missing_gcs_key_error,
-    stop_max_attempt_number=2,
-)
-def get_gs_files(
-    bucket: str, prefix: str, extensions: Sequence[str] = (".tif",)
-) -> List[str]:
-    """Get all geotiffs in GCS."""
-
-    try:
-        storage_client = storage.Client()
-    except DefaultCredentialsError:
-        raise MissingGCSKeyError()
-
-    blobs = storage_client.list_blobs(bucket, prefix=prefix)
-    files = [
-        f"/vsigs/{bucket}/{blob.name}"
-        for blob in blobs
-        if any(blob.name.endswith(ext) for ext in extensions)
-    ]
-    return files
+from gfw_pixetl.utils.aws import get_aws_files
+from gfw_pixetl.utils.google import get_gs_files
+from gfw_pixetl.utils.utils import DummyTile
 
 
 def get_key_from_vsi(vsi_path: str) -> str:
